CSS shenanigans
Triying to do some relatively advanced things with css to have a very neat essence skills style that can be expanded upon. It would probably be best to throw javascript into the mix, but that will be done in due time. I still have a lot to learn and remember.